Add 2 failing tests and remove a poorly written one.
test_deadline_cancel doesn't really work because it's super hard
to synchronize cone_deadline and cone_cancel so that they both trigger
while the coroutine is sleeping. Basically, there are two queues,
the "now" queue and the "later" queue, and the deadline is first
pushed onto the "later" queue. When it triggers, it is moved to
the end of the "now" queue, and when it is popped off, the coroutine
is actually marked as timed out and scheduled for execution.
Meanwhile, `cone_cancel` marks & schedules the coroutine directly,
pushing it onto the "now" queue. So to imitate simultaneity,
`cone_cancel` needs to be called after the deadline has moved
to the "now" queue, but before it is popped off.