Do not move timed events to the "now" queue.
That causes a window to occur when a timed event has not yet
happened, but can no longer be cancelled. If that event is a deadline
and the coroutine wakes up during the window, it will behave as if
the blocking call succeeded, but will fail at clearing the deadline
and will receive ETIMEDOUT at the next blocking call.