Allow disabling cancellation and deadlines.
Mostly useful when a function has shared its stack-allocated variables
with another coroutine and wants to make sure that it is dead before
returning by cancelling it and waiting for completion with interrupts
disabled. In this case, interrupting the wait will not lead to anything
good, only a use of stack after return.